refactor : auth 패키지 수정 및 검증로직 추가
