+++ src/main/java/com/nocaffeine/ssgclone/member/application/AuthService.java
@@ -1,17 +1,16 @@
 package com.nocaffeine.ssgclone.member.application;
 
-import com.nocaffeine.ssgclone.member.dto.request.AuthCheckRequestDto;
-import com.nocaffeine.ssgclone.member.dto.request.AuthEmailRequestDto;
-import com.nocaffeine.ssgclone.member.dto.request.SnsMemberAddRequestDto;
-import com.nocaffeine.ssgclone.member.dto.request.SnsMemberLoginRequestDto;
+import com.nocaffeine.ssgclone.member.dto.request.*;
 import com.nocaffeine.ssgclone.member.dto.response.TokenResponseDto;
 
 public interface AuthService {
     void snsAddMember(SnsMemberAddRequestDto snsMemberAddRequestDto);
     TokenResponseDto snsLogin(SnsMemberLoginRequestDto snsMemberLoginRequestDto);
 
+    void addMember(MemberSaveRequestDto memberSaveRequestDto);
+    void duplicationEmail(String email);
+    TokenResponseDto logIn(MemberLoginRequestDto memberLoginRequestDto);
 
     void emailAuth(AuthEmailRequestDto authEmailRequestDto);
-
-    void emailAuthCodeCheck(AuthCheckRequestDto authCheckRequestDto);
+    void emailAuthCodeCheck(String email, String code);
 }
