+++ src/main/java/com/nocaffeine/ssgclone/member/application/MemberService.java
@@ -7,10 +7,6 @@ import com.nocaffeine.ssgclone.member.dto.response.TokenResponseDto;
 
 public interface MemberService {
 
-    void duplicationEmail(String email);
-    void addMember(MemberSaveRequestDto memberSaveRequestDto);
-
-    TokenResponseDto logIn(MemberLoginRequestDto memberLoginRequestDto);
 
     void updatePassword(String memberUuid, MemberPasswordRequestDto memberPasswordRequestDto);
