+++ src/main/java/com/nocaffeine/ssgclone/member/vo/request/SnsMemberLoginRequestVo.java
@@ -5,4 +5,6 @@ import lombok.Getter;
 @Getter
 public class SnsMemberLoginRequestVo {
     private String snsId;
+    private String snsType;
+    private String email;
 }
